Crate gron

Convertor of JSON text into gron format

Gron is a representation that is easy to grep. Similarly to JSON it can be easily evaluated with javascript interpreter.

Original gron

Documentation

Traits

ToGron
An (internal) interface to gronify of json like data types

Functions

json_to_gron
Converts JSON structure from into gron format.